# tests/DecisionForge.ContractTests/Snapshots/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DecisionForge API",
    "description": "Secure same-origin API for explainable procurement decisions.",
    "version": "v1"
  },
  "paths": {
    "/version": {
      "get": {
        "tags": [
          "DecisionForge.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/antiforgery": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Issue a cookie-bound browser antiforgery token.",
        "operationId": "GetAntiforgeryToken",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AntiforgeryTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Create a secure cookie session.",
        "operationId": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          },
          "423": {
            "description": "Locked",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "End the current cookie session.",
        "operationId": "Logout",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Return the authenticated user's roles and permissions.",
        "operationId": "GetCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AntiforgeryTokenResponse": {
        "required": [
          "requestToken",
          "headerName"
        ],
        "type": "object",
        "properties": {
          "requestToken": {
            "type": "string"
          },
          "headerName": {
            "type": "string"
          }
        }
      },
      "ApiProblemDetails": {
        "required": [
          "type",
          "title",
          "status",
          "errorCode",
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "errorCode": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ApiValidationError"
            }
          }
        },
        "example": {
          "type": "https://decisionforge.local/problems/validation.field",
          "title": "The request contains invalid fields.",
          "status": 400,
          "errorCode": "validation.field",
          "traceId": "00-example-trace-00",
          "instance": "/api/v1/example"
        }
      },
      "ApiValidationError": {
        "required": [
          "code",
          "path",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "CurrentUserResponse": {
        "required": [
          "id",
          "email",
          "displayName",
          "roles",
          "permissions"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "example": {
          "email": "requester@decisionforge.local",
          "password": "configured-demo-password"
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "description": "HTTP-only secure same-origin application cookie.",
        "name": "__Host-DecisionForge-Auth",
        "in": "cookie"
      }
    }
  },
  "tags": [
    {
      "name": "DecisionForge.Api"
    },
    {
      "name": "Authentication"
    }
  ]
}